Added PacketHandler.cs
This is pretty much the bread and butter of the game server. It receives packets from the game client and decides what to do with them. 

Most of the packet structures have already been discovered for this patch version - but if we wanted to support a new packet (perhaps for a newer version of the game) - we'd just get the client to send over a packet and add a new case for the packet type (each packet has an ID), we'd then see what each part of the packet represents and make the server respond accordingly.